Create separate testing folder for io
